requirements.txt
python-telegram-bot==20.0a4
asyncio
pytest
pre-commit
pylint
